mini-cakestore-main/pyproject.toml
[project]
name = "mini-cakestore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "flask-bcrypt>=1.0.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
]